requirements.txt
absl-py==0.9.0
astunparse==1.6.3
bert-for-tf2==0.14.5
blinker==1.4
brotlipy==0.7.0
cachetools==4.1.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
Flask==1.1.2
gast==0.3.3
google-auth @ file:///tmp/build/80754af9/google-auth_1594357566944/work
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py @ file:///tmp/build/80754af9/h5py_1593454119955/work
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1594371495811/work
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1596233707986/work
oauthlib==3.1.0
opt-einsum==3.1.0
params-flow==0.8.2
protobuf==3.12.3
py-params==0.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
PyJWT==1.7.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
rsa==4.0
scipy @ file:///tmp/build/80754af9/scipy_1592930510235/work
six==1.15.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
tensorflow-hub==0.8.0
termcolor==1.1.0
tqdm==4.48.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
